requirements.txt
pytube 
yt-dlp
opencv-python 
torch 
torchvision 
transformers 
pandas 
numpy 
matplotlib 
albumentations 
scikit-learn 
mediapipe
ffmpeg-python